+++ src/main.js
@@ -5,6 +5,9 @@ import App from './components/App/App'
 
 /* eslint-disable no-unused-vars */
 
+import VueRouter from 'vue-router'
+Vue.use(VueRouter)
+
 // vue-material is used for md-tooltips, which display
 // calculator variable info
 import VueMaterial from 'vue-material'
@@ -48,12 +51,38 @@ Vue.component('calc-var-checkbox', CalcVarCheckbox)
 
 import store from './store'
 
+// 1. Define route components.
+// These can be imported from other files
+const Foo = { template: '<div>foo</div>' }
+const Bar = { template: '<div>bar</div>' }
+
+// 2. Define some routes
+// Each route should map to a component. The "component" can
+// either be an actual component constructor created via
+// Vue.extend(), or just a component options object.
+// We'll talk about nested routes later.
+const routes = [
+  { path: '/foo', component: Foo },
+  { path: '/bar', component: Bar }
+]
+
+// 3. Create the router instance and pass the `routes` option
+// You can pass in additional options here, but let's
+// keep it simple for now.
+const router = new VueRouter({
+  routes // short for routes: routes
+})
+
+import { sync } from 'vuex-router-sync'
+sync(store, router)
+
 // ROOT INSTANCE
 /* eslint-disable no-new */
 var vm = new Vue({
   el: '#app',
   store,
   template: '<App/>',
-  components: { App }
+  components: { App },
+  router
 })